a little bit of cleanup and reorganization
