--- backend/src/app.controller.ts
@@ -1,13 +1,0 @@
-import { Controller, Get, UseGuards } from '@nestjs/common';
-import { AppService } from './app.service';
-import { OauthGard42Guard } from './auth/guards/oauth-gard42.guard';
-
-@Controller()
-export class AppController {
-  constructor(private readonly appService: AppService) {}
-
-  @Get()
-  getHello(): string {
-    return this.appService.getHello();
-  }
-}
