--- backend/src/users/avatars.config.ts
@@ -1,18 +1,0 @@
-import { extname } from 'path';
-import { existsSync, mkdirSync } from 'fs';
-import { diskStorage } from 'multer';
-import { v4 as uuid } from 'uuid';
-import { HttpException, HttpStatus } from '@nestjs/common';
-
-export const avatarOptions = {
-	limits: { 
-		fileSize: 1024 * 1024 
-	},
-    fileFilter: (req: any, file: any, cb: any) => {
-        if (file.mimetype.match(/\/(jpg|jpeg|png|gif)$/)) {
-            cb(null, true);
-        } else {
-            cb(new HttpException(`Unsupported file type ${extname(file.originalname)}`, HttpStatus.BAD_REQUEST), false);
-        }
-    },
-};
