+++ backend/src/users/users.module.ts
@@ -1,13 +1,13 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UsersController } from './users.controller';
-import { UsersService } from './users.service';
+import { UsersService } from './services/users.service';
 import { User } from './entities/user.entity';
 import { PassportModule } from '@nestjs/passport';
 import { JwtModule } from '@nestjs/jwt';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { JwtStrategy } from 'src/users/strategy/jwt.strategy';
-import { AvatarsService } from 'src/users/avatars.service';
+import { AvatarsService } from 'src/users/services/avatars.service';
 import { Avatar } from './entities/avatar.entity';
 
 @Module({
