+++ backend/src/users/users.service.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { UsersService } from './users.service';
+import { UsersService } from './services/users.service';
 
 describe('UsersService', () => {
   let service: UsersService;
